Add handling of --extcap-version command line parameter
Since Wireshark 2.9 the extcap call to query its interface is extended
with the extcap-version parameter. The command line option parser is
not expecting this, emitting a message in stderr.
This change adds the extcap-version parameter to the expected set, to
stop the message being emitted on stderr, leaving out any processing
of the parameter value itself, which USBPcapCMD does not need, yet.