Fix full CPU core usage when using extcap API
Read from standard input only if it is valid. The significant CPU usage
was caused in the loop that waits for exit condition. The handle could not
be used in WaitForMultipleObjects() and thus it failed with INVALID HANDLE
error. As this was not handled, the code continued busy looping until
data->process was set to FALSE.

Fixes: #32