Rework GET DESCRIPTOR control transfers capture
Do not attempt to generate the Setup data when the IRP is travelling
from FDO to PDO. Instead, record the IRP information in URB IRP table.
When the IRP goes back from PDO to FDO, if the IRP function code changed
to control transfer, use the information from URB IRP table together
with the setup data entry from control transfer structure, to generate
the request pcap packet.

If the IRP function code does not denote control transfer and the IRP
was logged on the FDO to PDO route, log the packet as unknown URB.

Fixes #83