Restore control transfer specific URB handling
THe official USB 3.0 stack, unlike USB 2.0 stack, does not necessarily
change the URB function when the IRP is travelling back from PDO.

USBPcap 1.4.1.0 introduced new mechanism for control transfer capture
that relied on the (false) assumption that the IRP will have the URB
function changed to URB_FUNCTION_CONTROL_TRANSFER when travelling back
from PDO. This resulted in DEVICE and CONFIGURATION descriptors not being
captured for USB 3.0 devices connected to USB 3.0 host on recent Windows
versions.

This change restores the old mechanism for control transfer capture if
the IRP does not travel back with changed URB function code. As the
Setup packet is now manually created it should be checked if it is still
affected by the issue #83.